% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cr_token.R
\name{set_cr_token}
\alias{set_cr_token}
\title{Set Clash Royale API token}
\usage{
set_cr_token(token)
}
\arguments{
\item{token}{A character string containing the user's API token.}
}
\description{
Saves the user's Clash Royale API token to their .Renviron file so that
ClashR can automatically authenticate API requests.
}
\examples{
\dontrun{
  set_cr_token("YOUR_TOKEN_HERE")
}
}
